+++ CherryCrisis/CherryEngine/src/picking_renderpass.cpp
@@ -104,7 +104,8 @@ void PickingRenderPass::Execute(Framebuffer& fb, Viewer*& viewer)
 		glDrawElements(GL_TRIANGLES, gpuMesh->indicesCount, GL_UNSIGNED_INT, nullptr);
 	}
 
-	viewProjection = CCMaths::Matrix4::Orthographic(0.0f, fb.colorTex.width, 0.0f, fb.colorTex.height, -1.f, 5.f);
+	viewProjection = CCMaths::Matrix4::Orthographic(0.0f, static_cast<const float>(fb.colorTex.width),
+													0.0f, static_cast<const float>(fb.colorTex.height), -1.f, 5.f);
 	glUniformMatrix4fv(glGetUniformLocation(m_program->m_shaderProgram, "uViewProjection"), 1, GL_FALSE, viewProjection.data);
 
 	for (UIItem* uiRdr : m_uiRenderers)
